+++ utilities.cpp
@@ -72,7 +72,7 @@ template <typename T> bool zcmp(const Vec3<T> &u, const Vec3<T> &v) {
    return u.z < v.z;
 }
 
-bool(*func[3])(const Vec3f &, const Vec3f &) = {xcmp, ycmp, zcmp};
+bool(*func[3])(const Vec3d &, const Vec3d &) = {xcmp, ycmp, zcmp};
 
 template<typename T> void print(const Vec3<T> &u) {
    cout << u.x << ' ' << u.y << ' ' << u.z << ' ';
@@ -115,12 +115,12 @@ template<typename T> bool operator==(const Vec3<T> &u, const Vec3<T> &v) {
    return u.x == v.x && u.y == v.y && u.z == v.z;
 }
 
-template<typename T> float len(const Vec3<T> &u) {
+template<typename T> double len(const Vec3<T> &u) {
    return sqrt(u.x*u.x + u.y*u.y + u.z*u.z);
 }
 
 template<typename T> Vec3<T> unit(const Vec3<T> &u) {
-   float m = len(u);
+   double m = len(u);
    return Vec3<T>(u.x/m, u.y/m, u.z/m);
 }
 
@@ -138,10 +138,10 @@ template<typename T> Vec3<T> reflect(const Vec3<T> &u, const Vec3<T> &n) {
 }
 
 // UNTESTED
-template<typename T> Vec3<T> refract(const Vec3<T> &u, const Vec3<T> &n, const float r) {
-   float c1 = dot(u, n);
-   float e = 1.0/r;
+template<typename T> Vec3<T> refract(const Vec3<T> &u, const Vec3<T> &n, const double r) {
+   double c1 = dot(u, n);
+   double e = 1.0/r;
    if (c1<0) c1 = -c1; else e = -r;
-   float c2 = 1.0 - e*e*(1.0 - c1*c1);
-   return c2 < 0 ? Vec3f() : e*u + (e*c1 - sqrt(c2))*n;
+   double c2 = 1.0 - e*e*(1.0 - c1*c1);
+   return c2 < 0 ? Vec3d() : e*u + (e*c1 - sqrt(c2))*n;
 }
